+++ libraries/SysAPI/src/Process.cpp
@@ -165,6 +165,8 @@ Process::Process(const Arguments &arguments, const Environment &environment) {
     API_RETURN_ASSIGN_ERROR("no executable was found", EINVAL);
   }
 
+  bool is_process_launched = false;
+
 #if defined __win32
   m_process_information = new PROCESS_INFORMATION;
   *m_process_information = PROCESS_INFORMATION{};
@@ -175,11 +177,14 @@ Process::Process(const Arguments &arguments, const Environment &environment) {
   const var::PathString pwd = environment.find("PWD");
 
   int stdout_fd = dup(_fileno(stdout));
+  int stderr_fd = dup(_fileno(stderr));
 
   // change stdout and stderr to specify spawned process stdout, stderr
-  _dup2(m_pipe.write_file().fileno(), _fileno(stdout));
-  _dup2(_fileno(stdout), _fileno(stderr));
-  m_pipe.write_file() = fs::File();
+  _dup2(m_standard_output.pipe.write_file().fileno(), _fileno(stdout));
+  _dup2(m_standard_error.pipe.write_file().fileno(), _fileno(stderr));
+
+  m_standard_output.pipe.write_file() = fs::File();
+  m_standard_error.pipe.write_file() = fs::File();
 
   if (pwd.is_empty() == false) {
     _chdir(pwd.cstring());
@@ -195,6 +200,10 @@ Process::Process(const Arguments &arguments, const Environment &environment) {
   _dup2(stdout_fd, _fileno(stdout));
   _close(stdout_fd);
 
+  // restore stderr
+  _dup2(stderr_fd, _fileno(stderr));
+  _close(stderr_fd);
+
   if (pwd.is_empty() == false) {
     _chdir(cwd.cstring());
   }
@@ -205,7 +214,16 @@ Process::Process(const Arguments &arguments, const Environment &environment) {
     API_RETURN_ASSIGN_ERROR("failed to spawn", errno);
   }
 
+  is_process_launched = true;
+
 #else
+
+  const auto pwd = environment.find("PWD");
+  if (fs::FileSystem().directory_exists(pwd) == false) {
+    API_RETURN_ASSIGN_ERROR(
+      ("cannot change dir to `" | var::StringView(pwd)) | "`",
+      errno);
+  }
 
   m_pid = API_SYSTEM_CALL("fork()", fork());
   if (m_pid == 0) {
@@ -217,22 +235,34 @@ Process::Process(const Arguments &arguments, const Environment &environment) {
       API_RETURN_ASSIGN_ERROR("failed to chdir to PWD", errno);
     }
 
-    dup2(m_pipe_output.write_file().fileno(), STDOUT_FILENO);
+    dup2(m_standard_output.pipe.write_file().fileno(), STDOUT_FILENO);
     // stdout will now write to the pipe -- this fileno isn't needed anymore
     // but doesn't necessarily have to be closed
-    m_pipe_output.write_file() = fs::File();
+    m_standard_output.pipe.write_file() = fs::File();
 
-    dup2(m_pipe_error.write_file().fileno(), STDERR_FILENO);
-    // stdout will now write to the pipe -- this fileno isn't needed anymore
+    dup2(m_standard_error.pipe.write_file().fileno(), STDERR_FILENO);
+    // stderr will now write to the pipe -- this fileno isn't needed anymore
     // but doesn't necessarily have to be closed
-    m_pipe_error.write_file() = fs::File();
+    m_standard_error.pipe.write_file() = fs::File();
 
     // replace the current process with the one specified
     ::execve(args.path(), args.m_arguments.data(), environ);
     perror("failed to launch\n");
     exit(1);
   }
+
+  is_process_launched = m_pid > 0;
 #endif
+
+  if (is_process_launched) {
+    m_standard_output_redirect_options
+      = {.redirect = &m_standard_output, .self = this};
+    m_standard_error_redirect_options
+      = {.redirect = &m_standard_error, .self = this};
+
+    m_standard_output.start_thread(&m_standard_output_redirect_options);
+    m_standard_error.start_thread(&m_standard_error_redirect_options);
+  }
 }
 
 Process::~Process() {
@@ -254,7 +284,8 @@ Process &Process::wait() {
 
   m_process = INVALID_HANDLE_VALUE;
   m_status = int(code);
-
+  m_standard_error.wait_stop();
+  m_standard_output.wait_stop();
   return *this;
 
 #else
@@ -268,6 +299,8 @@ Process &Process::wait() {
   if (result == m_pid) {
     m_pid = -1;
     m_status = status;
+    m_standard_error.wait_stop();
+    m_standard_output.wait_stop();
   }
 #endif
 
@@ -292,6 +325,8 @@ bool Process::is_running() {
   }
 
   m_status = int(code);
+  m_standard_error.wait_stop();
+  m_standard_output.wait_stop();
   return false;
 #else
   if (m_pid < 0) {
@@ -305,6 +340,9 @@ bool Process::is_running() {
   if (result == m_pid) {
     m_status = status;
     m_pid = -1;
+
+    m_standard_error.wait_stop();
+    m_standard_output.wait_stop();
     return false;
   }
 #endif
@@ -312,36 +350,99 @@ bool Process::is_running() {
   return true;
 }
 
-Process &Process::read_output() {
-  auto read_pipe = [&](const fs::FileObject &data, const fs::FileObject &pipe) {
-    var::Array<char, 2048> buffer;
-    api::ErrorScope error_scope;
-    int bytes_read = 0;
-    do {
-      bytes_read = pipe.read(buffer).return_value();
-      if (bytes_read > 0) {
-        data.write(var::View(buffer.data(), bytes_read));
-      }
-    } while (bytes_read > 0);
-  };
-
-  read_pipe(m_standard_output, m_pipe_output.read_file());
-  read_pipe(m_standard_error, m_pipe_error.read_file());
-  return *this;
-}
-
 var::String Process::get_standard_output() {
-  read_output();
+  thread::Mutex::Scope mutex_scope(m_standard_output.mutex);
   return var::String(var::StringView(
-    reinterpret_cast<const char*>(m_standard_output.data().data_u8()),
-    m_standard_output.data().size()));
+    reinterpret_cast<const char *>(
+      m_standard_output.data_file.data().data_u8()),
+    m_standard_output.data_file.data().size()));
 }
 
 var::String Process::get_standard_error() {
-  read_output();
+  thread::Mutex::Scope mutex_scope(m_standard_error.mutex);
   return var::String(var::StringView(
-    reinterpret_cast<const char*>(m_standard_error.data().data_u8()),
-    m_standard_error.data().size()));
+    reinterpret_cast<const char *>(m_standard_error.data_file.data().data_u8()),
+    m_standard_error.data_file.data().size()));
+}
+
+var::String Process::read_standard_output() { return m_standard_output.read(); }
+
+var::String Process::read_standard_error() { return m_standard_error.read(); }
+
+void Process::update_redirect(const RedirectOptions *options) {
+  // reads the pipe in a thread
+#if defined __win32
+  options->redirect->thread_handle = GetCurrentThread();
+#endif
+  var::Array<char, 2048> buffer;
+  auto &data_file = options->redirect->data_file;
+  auto &pipe = options->redirect->pipe;
+  auto &mutex = options->redirect->mutex;
+  data_file.seek(0, fs::File::Whence::end);
+  bool is_stop_requested = options->redirect->is_stop_requested;
+  do {
+    api::ErrorScope error_scope;
+    const auto bytes_read = pipe.read_file().read(buffer).return_value();
+    {
+      thread::Mutex::Scope mutex_scope(mutex);
+      if (bytes_read > 0) {
+        // data file is write append, we want to preserve
+        // the file location for reading the data file
+        fs::File::LocationScope location_scope(data_file);
+        data_file.write(var::View(buffer.data(), bytes_read));
+      }
+      is_stop_requested = options->redirect->is_stop_requested;
+
+    }
+  } while (!is_stop_requested);
+}
+
+void *Process::update_redirect_thread_function(void *args) {
+  const auto options = reinterpret_cast<const RedirectOptions *>(args);
+  options->self->update_redirect(options);
+  return nullptr;
+}
+
+void Process::Redirect::start_thread(Process::RedirectOptions *options) {
+  thread = thread::Thread(
+    thread::Thread::Attributes().set_joinable(),
+    options,
+    update_redirect_thread_function);
+}
+
+var::String Process::Redirect::read() {
+  thread::Mutex::Scope ms(mutex);
+  var::Array<char, 1024> buffer;
+  var::String result;
+  int bytes_read = 0;
+  do {
+    api::ErrorScope error_scope;
+    bytes_read = data_file.read(var::View(buffer.data(), buffer.count() - 1))
+                   .return_value();
+    if (bytes_read > 0) {
+      buffer.at(bytes_read) = 0;
+      result += buffer.data();
+    }
+  } while (bytes_read > 0);
+  return result;
+}
+
+#if defined __win32
+#undef _WIN32_WINNT
+#define _WIN32_WINNT 0x0600
+#include <ioapiset.h>
+#endif
+
+void Process::Redirect::wait_stop() {
+  {
+    thread::Mutex::Scope ms(mutex);
+    is_stop_requested = true;
+  }
+
+  thread.cancel();
+  if( thread.is_joinable() ){
+    thread.join();
+  }
 }
 
 #endif